Creata la GUI, parte di sinistra funzionante

+++ src/core/Utils.java
@@ -4,24 +4,24 @@ import java.io.*;
 import java.nio.file.*;
 
 public class Utils {
-    public static int nPartsCalc(File f, int dim){
+    public static int nPartsCalc(File f, long dim){
         int lenght = (int) f.length();
 
         //Calcola il numero di parti
-        int partNum = lenght / dim;
+        long partNum = lenght / dim;
         if (partNum * dim < lenght) partNum++;
 
-        return partNum;
+        return (int) partNum;
     }
 
-    public static int dimCalc(File f, int nParts){
-        int lenght = (int) f.length();
+    public static long dimCalc(File f, int nParts){
+        long lenght = f.length();
 
         // Calcola la dimensione di una parte
-        double dim = (double)lenght / nParts;
-        if ((int)dim * nParts != lenght) dim++;
+        long dim = lenght / nParts;
+        if (dim * nParts != lenght) dim++;
 
-        return (int)dim;
+        return dim;
     }
 
     public static String getDirectory(File f){
